Fix deletion with new ID scheme, bug fix map_date

+++ harvester/solr_updater.py
@@ -39,13 +39,15 @@ def make_datetime(dstring):
     Right now formats are YYYY or YYYY-MM-DD
     '''
     dt = None
+    if not dstring:
+        return dt
     try:
         dint = int(dstring)
         dt = datetime.datetime(dint, 1, 1)
     except ValueError:
         pass
     except TypeError, e:
-        print >> sys.stderr, 'Date type err:{} {}'.format(dstring, e)
+        print >> sys.stderr, 'Date type err DATA:{} ERROR:{}'.format(dstring, e)
     try:
         strfmt = '%Y-%m-%d'
         dt = datetime.datetime.strptime(dstring, strfmt)
@@ -86,23 +88,29 @@ def map_date(d):
             try:
                 displayDate, dt_start, dt_end = get_dates_from_date_obj(date_source)
                 dates.append(displayDate)
-                dates_start.append(dt_start)
-                dates_end.append(dt_end)
+                if dt_start:
+                     dates_start.append(dt_start)
+                if dt_end:
+                    dates_end.append(dt_end)
             except KeyError:
                 pass
         else: #should be list
             for dt in date_source:
                 displayDate, dt_start, dt_end = get_dates_from_date_obj(dt)
                 dates.append(displayDate)
-                dates_start.append(dt_start)
-                dates_end.append(dt_end)
+                if dt_start:
+                     dates_start.append(dt_start)
+                if dt_end:
+                    dates_end.append(dt_end)
 
     date_map['date'] = dates
 
     dates_start = sorted(dates_start)
     dates_end = sorted(dates_end)
-    start_date = dates_start[0]
-    end_date = dates_end[0]
+    if len(dates_start):
+    	start_date = dates_start[0]
+    if len(dates_end):
+    	end_date = dates_end[0]
     # fill in start_date == end_date if only one exists
     start_date = end_date if not start_date else start_date
     end_date = start_date if not end_date else end_date
@@ -472,10 +480,14 @@ def main(url_couchdb=None, dbname=None, url_solr=None, all_docs=False, since=Non
         if row.get('deleted', False):
             #need to get the solr doc for this couch
             #TODO: test for this?
-            sdoc = solr_db.select(q='harvest_id_s:cur_id')
-            print('====DELETING: {0} -- {1}'.format(cur_id, sdoc['id']))
-            solr_db.delete(id=sdoc['id'])
-            n_delete += 1
+            resp = solr_db.select(q=''.join(('harvest_id_s:"', cur_id, '"')))
+            if resp.numFound == 1:
+                sdoc = resp.results[0]
+            	print('====DELETING: {0} -- {1}'.format(cur_id, sdoc['id']))
+            	solr_db.delete(id=sdoc['id'])
+            	n_delete += 1
+            else:
+                print("-----DELETION of {} - FOUND {} docs".format(cur_id, resp.numFound))
         else:
             doc = db.get(cur_id)
             try:
